Lazy-load the relationship before getting start and end nodes or the type.
These values wouldn't be set on a load or update, and if they were, they don't get updated anyway.  This allows instantiating a relationship and setting its id (say from a path-find or listing all relationships on a node) without having to get the relationship data from the server until absolutely necessary.